[reorg] Move ReverseChainSingleSubst to new layout.
