+++ src/Makefile.sources
@@ -102,6 +102,8 @@ HB_BASE_sources = \
 	OT/Layout/GSUB/LigatureSubst.hh \
 	OT/Layout/GSUB/LigatureSet.hh \
 	OT/Layout/GSUB/Ligature.hh \
+	OT/Layout/GSUB/ReverseChainSingleSubst.hh \
+	OT/Layout/GSUB/ReverseChainSingleSubstFormat1.hh \
 	hb-ot-layout-gsubgpos.hh \
 	hb-ot-layout-jstf-table.hh \
 	hb-ot-layout.cc \
